line-bot-lambda/requirements.txt
line-bot-sdk>=3.14.0
boto3>=1.42.0
